+++ src/application/useCases/getRecipientNotifications.ts
@@ -1,0 +1,26 @@
+import { Notification } from '@application/entities/notification';
+import { NotificationsRepository } from '@application/repositories/notificationsRepository';
+import { Injectable } from '@nestjs/common';
+
+interface GetRecipientNotificationsRequest {
+  recipientId: string;
+}
+
+interface GetRecipientNotificationsResponse {
+  notifications: Notification[];
+}
+
+@Injectable()
+export class GetRecipientNotifications {
+  constructor(private notificationRepository: NotificationsRepository) {}
+
+  async execute(
+    request: GetRecipientNotificationsRequest,
+  ): Promise<GetRecipientNotificationsResponse> {
+    const { recipientId } = request;
+    const notifications =
+      await this.notificationRepository.findManyByRecipientId(recipientId);
+
+    return { notifications };
+  }
+}