+++ src/application/useCases/readNotification.spec.ts
@@ -1,0 +1,36 @@
+import { makeNotification } from '@test/factories/notificationFactory';
+import { InMemoryNotificationsRepository } from '@test/repositories/inMemoryNotificationsRepository';
+import { NotificationNotFound } from './errors/notificationNotFoundError';
+import { ReadNotification } from './readNotification';
+
+describe('Read notification', () => {
+  it('should be able to read a notification', async () => {
+    const notificationRepository = new InMemoryNotificationsRepository();
+    const readNotification = new ReadNotification(notificationRepository);
+
+    const notification = makeNotification();
+
+    await notificationRepository.create(notification);
+
+    await readNotification.execute({ notificationId: notification.id });
+
+    expect(notificationRepository.notifications).toHaveLength(1);
+    expect(notificationRepository.notifications[0].readAt).toEqual(
+      expect.any(Date),
+    );
+  });
+
+  it('should not be able to read a non existing notification', async () => {
+    const notificationRepository = new InMemoryNotificationsRepository();
+    const readNotification = new ReadNotification(notificationRepository);
+
+    await notificationRepository.create(makeNotification());
+
+    expect(notificationRepository.notifications).toHaveLength(1);
+    expect(() => {
+      return readNotification.execute({
+        notificationId: 'id-not-in-database',
+      });
+    }).rejects.toThrow(NotificationNotFound);
+  });
+});