+++ src/infra/database/prisma/repositories/prismaNotificationsRepository.ts
@@ -1,22 +1,68 @@
 import { Injectable } from '@nestjs/common';
-import { Notification } from '../../../../application/entities/notification';
-import { NotificationsRepository } from '../../../../application/repositories/notificationsRepository';
-import { PrismaService } from '../../prisma.service';
+import { Notification } from '@application/entities/notification';
+import { NotificationsRepository } from '@application/repositories/notificationsRepository';
+import { PrismaService } from '../prisma.service';
+import { PrismaNotificationMapper } from '../mappers/prismaNotificationMapper';
 
 @Injectable()
 export class PrismaNotificationRepository implements NotificationsRepository {
-  constructor(private prismaService: PrismaService) {}
+  constructor(private prisma: PrismaService) {}
+
+  async findById(notificationId: string): Promise<Notification> {
+    const notification = await this.prisma.notification.findUnique({
+      where: {
+        id: notificationId,
+      },
+    });
+
+    return PrismaNotificationMapper.toDomain(notification);
+  }
+
+  async findManyByRecipientId(recipientId: string): Promise<Notification[]> {
+    const rawNotifications = await this.prisma.notification.findMany({
+      where: {
+        recipientId,
+      },
+    });
+
+    //return notifications = rawNotifications.forEach((raw) =>
+    //  PrismaNotificationMapper.toDomain(raw),
+    // );
+
+    //return rawNotifications.map((raw) => {
+    //  return PrismaNotificationMapper.toDomain(raw);
+    //});
+
+    return rawNotifications.map(PrismaNotificationMapper.toDomain);
+  }
+
+  async countManyByRecipientId(recipientId: string): Promise<number> {
+    const count = await this.prisma.notification.count({
+      where: {
+        recipientId: recipientId,
+      },
+    });
+
+    return count;
+  }
 
   async create(notification: Notification): Promise<void> {
-    await this.prismaService.notification.create({
-      data: {
-        id: notification.id,
-        content: notification.content.value,
-        category: notification.category,
-        recepientId: notification.recepientId,
-        readAt: notification.readAt,
-        createdAt: notification.createdAt,
+    const raw = PrismaNotificationMapper.toPrisma(notification);
+
+    await this.prisma.notification.create({
+      data: raw,
+    });
+  }
+
+  async save(notification: Notification): Promise<void> {
+    const raw = PrismaNotificationMapper.toPrisma(notification);
+    console.log(raw);
+
+    await this.prisma.notification.update({
+      where: {
+        id: raw.id,
       },
+      data: raw,
     });
   }
 }